Setup GPIO pins as TDD trigger signal
Pin 0 is used to trigger the PA to switch to TX mode. We use timed commands
as the ATR of the USRP was not applicable because the B200 requires a too
large burst_preamble. Right now we set the trigger when the frame starts,
maybe we need to start a bit earlier. The end of the trigger signal occurs
0.05 ms before the end of the subframe, this is feasible, as the last ofdm-
symbol of a sidelink subframe is ignored anyway.
Further, PIN 1 has been configured to be in an always on state.